backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.42.0
langchain>=0.3.14
langchain-anthropic>=0.3.4
langchain-google-genai>=2.1.0,<3.0
langchain-groq>=0.3.0,<1.0
langchain-core>=0.3.73,<1.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
tree-sitter-java>=0.23.0
tree-sitter-cpp>=0.23.0
tree-sitter-go>=0.23.0
tree-sitter-rust>=0.23.0
httpx>=0.27.0
watchdog>=4.0.0
